Code cleaning, refactor, unify conventions

+++ Assembly-CSharp/Global/BGOVERLAY_DEF.cs
@@ -9,9 +9,9 @@ public partial class BGOVERLAY_DEF
 	{
 		this.startOffset = 0L;
 		this.spriteList = new List<BGSPRITE_LOC_DEF>();
-        this.canCombine = true;
-        this.isCreated = false;
-    }
+		this.canCombine = true;
+		this.isCreated = false;
+	}
 
 	public void SetFlags(OVERLAY_FLAG flagDiff, Boolean isSet)
 	{
@@ -51,8 +51,8 @@ public partial class BGOVERLAY_DEF
 		this.scrY = reader.ReadInt16();
 		this.ParallaxDepthX = reader.ReadInt16();
 		this.ParallaxDepthY = reader.ReadInt16();
-		this.fracX = (float)reader.ReadInt16();
-		this.fracY = (float)reader.ReadInt16();
+		this.fracX = reader.ReadInt16();
+		this.fracY = reader.ReadInt16();
 		Byte bitPos = 0;
 		buffer = reader.ReadUInt32();
 		this.camNdx = (Byte)BitUtil.ReadBits(buffer, ref bitPos, 8);
@@ -89,12 +89,12 @@ public partial class BGOVERLAY_DEF
 	public Int16 ParallaxDepthX; // negative is slower (further) than main, positive faster (closer)
 	public Int16 ParallaxDepthY;
 
-	public float fracX;
-	public float fracY;
+	public Int16 fracX;
+	public Int16 fracY;
 
 	public Byte camNdx;
 	public Byte isXOffset;
-    public UInt32 indnum;
+	public UInt32 indnum;
 	public Byte viewportNdx;
 
 	public UInt16 spriteCount;
@@ -111,8 +111,8 @@ public partial class BGOVERLAY_DEF
 
 	public Transform transform;
 
-    public Boolean canCombine;
-    public Boolean isCreated;
+	public Boolean canCombine;
+	public Boolean isCreated;
 
 	public Boolean isMemoria = false;
 	public Vector2 memoriaSize;
